Commit changes for some of the review comments.
