+++ koku/masu/test/external/downloader/ocp/test_ocp_report_downloader.py
@@ -165,16 +165,6 @@ class OCPReportDownloaderTest(MasuTestCase):
             # Re-enable log suppression
             logging.disable(logging.CRITICAL)
 
-    def test_detect_type(self):
-        """Test the detect_type method."""
-        mock_df = Mock(columns=["pod_labels"])
-        report_type = self.ocp_report_downloader.detect_type(mock_df)
-        self.assertEqual(report_type, "pod_usage")
-
-        mock_df = Mock(columns=["foo"])
-        report_type = self.ocp_report_downloader.detect_type(mock_df)
-        self.assertEqual(report_type, None)
-
     def test_divide_csv_daily(self):
         """Test the divide_csv_daily method."""
 
@@ -182,16 +172,20 @@ class OCPReportDownloaderTest(MasuTestCase):
             filename = "storage_data.csv"
             file_path = f"{td}/{filename}"
             with patch("masu.external.downloader.ocp.ocp_report_downloader.pd") as mock_pd:
-                mock_report = {
-                    "interval_start": ["2020-01-01 00:00:00 +UTC", "2020-01-02 00:00:00 +UTC"],
-                    "persistentvolumeclaim_labels": ["label1", "label2"],
-                }
-                df = pd.DataFrame(data=mock_report)
-                mock_pd.read_csv.return_value = df
-                daily_files = self.ocp_report_downloader.divide_csv_daily(file_path, filename)
-                self.assertNotEqual([], daily_files)
-                self.assertEqual(len(daily_files), 2)
-                gen_files = ["storage_usage.2020-01-01.csv", "storage_usage.2020-01-02.csv"]
-                expected = [{"filename": gen_file, "filepath": f"{td}/{gen_file}"} for gen_file in gen_files]
-                for expected_item in expected:
-                    self.assertIn(expected_item, daily_files)
+                with patch(
+                    "masu.external.downloader.ocp.ocp_report_downloader.utils.detect_type",
+                    return_value=("storage_usage", None),
+                ):
+                    mock_report = {
+                        "interval_start": ["2020-01-01 00:00:00 +UTC", "2020-01-02 00:00:00 +UTC"],
+                        "persistentvolumeclaim_labels": ["label1", "label2"],
+                    }
+                    df = pd.DataFrame(data=mock_report)
+                    mock_pd.read_csv.return_value = df
+                    daily_files = self.ocp_report_downloader.divide_csv_daily(file_path, filename)
+                    self.assertNotEqual([], daily_files)
+                    self.assertEqual(len(daily_files), 2)
+                    gen_files = ["storage_usage.2020-01-01.csv", "storage_usage.2020-01-02.csv"]
+                    expected = [{"filename": gen_file, "filepath": f"{td}/{gen_file}"} for gen_file in gen_files]
+                    for expected_item in expected:
+                        self.assertIn(expected_item, daily_files)
